gyp: add a binding.gyp file for node-gyp support
Essentially the same setup as the `wscript`. It uses the `xml2-config`
command to get the necessary -l and -I flags to link to the system
libxml2.

Therefore, there's no Windows support yet. That will require bundling
libxml2 for Windows and statically linking to that, which is something
for down the road.