Support XPath expressions that yield attributes
XPath expressions like '@attr' evaluate to a nodeset containing
attributes.  This would sometimes work in libxmljs (when the libxml
xmlNode for an attribute already had a _private pointing to a libxmljs
attribute wrapper).  But in other cases you could end up with a libxmljs
element wrapper wrapping a libxmljs xmlAttr.  This makes it work in
general.