Test exposing SEGFAULT in XmlNode destructor
When you remove a node from the XmlTree, libxml removes that node and
all of its children from the tree, but doesn't free that memory.
However, these libxml nodes maybe referenced by libxmljs XmlNode's.  So,
when those  XmlNodes' destructors are called, it causes us to try to
doulbe free that memory.