GH-172 - SetDtd on XmlDocument
This should allow a user to set the doctype on a document. Internally
uses the createIntSubset function and applies the dtd to the document
passed into the function.

There were some issues relating to removal of the dtd by calling
xmlCreateIntSubset so instead we first have to get the dtd node, and
then basically remove it and free it before we can apply the new dtd.
This might be better as a few split api's rather than just one lump api
but I figured it would be a little easier this way.

Signed-off-by: Nick Campbell <nicholas.j.campbell@gmail.com>