Implement thread-local memory adjustments.
If the current thread is not the main V8 event thread, we don't report
memory adjustments back to V8 immediately, but instead keep track of all the
changes in a thread-local variable.  The net change will be reported back
after the worker is done.

All state changes are encapsulated in two RAII sentinels: WorkerParent and
WorkerSentinel.  The former is constructed and destroyed on the V8 thread,
the latter on the worker thread.  Having these in place helps to make client
code thread-safe with respect to the state in question.  It also provides a
convenient extension point: if, at some point, we need more code for
handling worker threads, all we have to do is extend these two classes.

The current delegation to the libxml2 memory management functions has ben
completely dropped.  The main use case for these would have been memory
debugging, so they do a lot of work (like maintaining linked lists between
allocated blocks) that we don't need.  The main thing we could have gained
from them, namely the call to xmlMemUsed, was not thread-safe in any case.