Complain when allocating without V8 instance.
This helps ensure that a worker will actually use the WorkerParent and
WorkerSentinel machinery.  So far, any memory handling on a worker which
failed to do so would fall under the clause where there is no current
instance.  That might cause memory adjustments to go unnoticed.  So it's
better to force users to update their code, to properly set up workers.