Get Travis to build for iojs as well
The problem with node-gyp not working with iojs can be solved by using the
version shipped with iojs, not the version from the latest npm we install
separately.  See https://github.com/nodejs/io.js/issues/433 for details.
So we update npm to latest upstream only for old versions of node.js,
not for the current version or for iojs.