Fix resolving flags in parseHtmlAsync and parseXmlAsync
The options that were passed to the `parseXmlAsync` and `parseHtmlAsync` method were not used to create the correct flags.

The XML or HTML options were not being processed in the parseBufferAsync method as they were in the `parseXml` or `parseHtml` methods.

Added a new method that resolves the options by the `FROM_BUFFER_ASYNC_TYPE`.

Deprecated the `parseBufferAsync` method.